add: metaportfolio
Add the class MetaPortfolio creating the meta ptf from a given analyst book and analyst predicted forward pnl.
This include two methods of portfolio construction :
- The "paper" way : by normalizing the signal on each side (long and short) so that te portfolio is self financing
- The common way by doing the aggregation directly at stock level and then normalizing the weights if needed
Also includes a method to transform the predicted pnl to a more suitable signal by increasing the amplitude of extremes (positive and negative forward pnl) and setting the signal with low applitude to 0 or very close.